ingest/requirements.txt
openai==1.66.3
telethon==1.39.0
supabase==2.13.0
python-dotenv==1.0.1
Pillow==11.1.0
python-slugify==8.0.4
